feedback

+++ backend/models/customer_feedback.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, DateTime, Numeric, Text, Boolean, ForeignKey
+from sqlalchemy import Column, String, Integer, DateTime, Text, Boolean, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
@@ -13,8 +13,8 @@ class DBCustomerFeedback(Base):
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     order_id = Column(Integer, ForeignKey('orders.id'), nullable=False)
-    restaurant_rating = Column(Numeric(3, 2), default=0.00, nullable=False)
-    delivery_rating = Column(Numeric(3, 2), default=0.00, nullable=False)
+    restaurant_rating = Column(Integer, default=0, nullable=False)
+    delivery_rating = Column(Integer, default=0, nullable=False)
     feedback = Column(Text, nullable=True)
     timestamp = Column(DateTime, default=datetime.now(local_tz), nullable=False)
     would_recommend = Column(Boolean, default=False, nullable=True)
