+++ backend/schemas/customer_feedback.py
@@ -7,13 +7,13 @@ local_tz = timezone(timedelta(hours=2))
 
 
 class CustomerFeedbackBase(BaseModel):
-    restaurant_rating: float = Field(
-        ..., ge=0.0, le=5.0, description="Rating must be between 0 and 5"
+    restaurant_rating: int = Field(
+        ..., ge=0, le=5, description="Rating must be between 0 and 5"
     )
-    delivery_rating: float = Field(
-        ..., ge=0.0, le=5.0, description="Rating must be between 0 and 5"
+    delivery_rating: int = Field(
+        ..., ge=0, le=5, description="Rating must be between 0 and 5"
     )
-    feedback: str = None
+    feedback: str = None 
     timestamp: datetime = datetime.now(local_tz)
     would_recommend: bool = False
